doDelete, doModify 권한체크. ResultData에 data1Name 추가

+++ src/main/java/com/cjh/exam/demo/controller/UsrArticleController.java
@@ -46,7 +46,7 @@ public class UsrArticleController {
 		
 		Article article = articleService.getArticle((int) writeArticleRd.getData1());
 		
-		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), article);
+		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), "article", article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
@@ -54,7 +54,7 @@ public class UsrArticleController {
 	public ResultData<List<Article>> getArticles() {
 
 		List<Article> articles = articleService.getArticles();
-		return ResultData.from("S-1", "게시물 리스트", articles);
+		return ResultData.from("S-1", "게시물 리스트", "articles", articles);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
@@ -65,24 +65,32 @@ public class UsrArticleController {
 			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
 
+		int loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Utility.f("%d번 게시물은 존재하지 않습니다", id));
 		}
+		
+		if(loginedMemberId != article.getMemberId()) {
+			return ResultData.from("F-B", "해당 게시물에 대한 권한이 없습니다");
+		}
 
 		articleService.deleteArticle(id);
 
-		return ResultData.from("S-1", Utility.f("%d번 게시물을 삭제했습니다", id), id);
+		return ResultData.from("S-1", Utility.f("%d번 게시물을 삭제했습니다", id), "id", id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(HttpSession httpSession, int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
 
 		if(httpSession.getAttribute("loginedMemberId") == null) {
 			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
+		
+		int loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		
 		Article article = articleService.getArticle(id);
 
@@ -90,9 +98,13 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Utility.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 
-		articleService.modifyArticle(id, title, body);
+		ResultData actorCanModifyRd = articleService.actorCanModify(loginedMemberId, article);
+		
+		if(actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
 
-		return ResultData.from("S-1", Utility.f("%d번 게시물을 수정했습니다", id), id);
+		return articleService.modifyArticle(id, title, body);
 	}
 
 	@RequestMapping("/usr/article/getArticle")
@@ -105,7 +117,7 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Utility.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 
-		return ResultData.from("S-1", Utility.f("%d번 게시물 입니다", id), article);
+		return ResultData.from("S-1", Utility.f("%d번 게시물 입니다", id), "article", article);
 	}
 
 }
